planes: Add planes dependency
This code depends on planes, which is a unofficial pygame package
that could be difficult to locate in the correct version.  For example,
this code is already somewhat depreciated, because for this
re-organization, I had to download the newest version of planes, which
uses a different font system that made all the font sizes wrong and
required editing of a lot of code.  So hopefully putting this in
the repo will prevent future issues